- I broke the build :(

+++ Vha.Common/Binary.cs
@@ -185,7 +185,7 @@ namespace Vha.Common
             if (length > UInt16.MaxValue)
                 throw new ArgumentException("string byte length exceeds " + UInt16.MaxValue);
             byte[] bytes = new byte[length + 2];
-            WriteUInt16(length, endianness).CopyTo(bytes, 0);
+            WriteUInt16((UInt16)length, endianness).CopyTo(bytes, 0);
             encoding.GetBytes(value, 0, value.Length, bytes, 2);
             return bytes;
         }
